Introducing align sort: sorting entities to maximize alignment of cpus on one entity
CPU allocation goals 1) Fragmentation avoidance 2) alignment
Today CPU are designed such that they work on fragementation avoidance.
There is no mechanism to enforce alignment.
With this change, we introduce flag alignSort ie
Entities are sorted in such a way that alignment is primary goal and
fragmentation secondary.
Entities are sorted such that first entites are those in which numCPU
can completely fit and then sorrted such that first we consume in order
of minimum number of cpus